first version of the nnpdf data package, with versioning
add utility function to read metadata just from dataset name

deprecate a bunch of functions

fix include
