+++ validphys2/src/validphys/tests/test_commondataparser.py
@@ -2,8 +2,8 @@ import numpy as np
 import pandas as pd
 import pytest
 
+from nnpdf_data.commondataparser import load_commondata
 from validphys.api import API
-from validphys.commondataparser import load_commondata
 from validphys.covmats_utils import construct_covmat
 from validphys.loader import FallbackLoader as Loader
 from validphys.tests.conftest import FIT, THEORYID_NEW
@@ -12,7 +12,7 @@ from validphys.tests.conftest import FIT, THEORYID_NEW
 def test_basic_commondata_loading():
     l = Loader()
     cd = l.check_commondata(setname="SLAC_NC_NOTFIXED_D_EM-F2", variant="legacy_dw")
-    res = load_commondata(cd)
+    res = load_commondata(cd.metadata)
     # Test commondata loading
     assert res.ndata == 211
     assert isinstance(res.commondata_table, pd.DataFrame)
@@ -30,7 +30,7 @@ def test_basic_commondata_loading():
     emptysyscd = l.check_posset(
         theoryID=THEORYID_NEW, setname='NNPDF_POS_2P24GEV_XDQ', postlambda=1e-10, rules=rules
     )
-    emptysysres = load_commondata(emptysyscd.commondata)
+    emptysysres = load_commondata(emptysyscd.commondata.metadata)
     assert emptysysres.nsys == 0
     assert emptysysres.systype_table.empty is True
 
@@ -40,7 +40,7 @@ def test_commondata_with_cuts():
     setname = "NMC_NC_NOTFIXED_P_EM-SIGMARED"
 
     cd = l.check_commondata(setname=setname, variant="legacy")
-    loaded_cd = load_commondata(cd)
+    loaded_cd = load_commondata(cd.metadata)
 
     fit_cuts = l.check_fit_cuts(fit=FIT, commondata=cd)
     internal_cuts = l.check_internal_cuts(cd, API.rules(theoryid=THEORYID_NEW, use_cuts="internal"))
@@ -110,9 +110,10 @@ def test_commondata_load_write_load(tmp):
     original_covmat = construct_covmat(original_stats, original_data.systematic_errors(fake_data))
     np.testing.assert_allclose(new_covmat, original_covmat)
 
+
 def test_variant_nnpdf_metadata():
     """Tests the undocumented feature of a variant which updates the key `experiment`
-          within the nnpdf_metadata
+    within the nnpdf_metadata
     """
     l = Loader()
     set_name = "SLAC_NC_NOTFIXED_D_EM-F2"
@@ -124,7 +125,9 @@ def test_variant_nnpdf_metadata():
         pcd2 = cd2.metadata.plotting_options
 
         # ensure the nnpdf_metadata and the plotting are changed
-        assert cd1.metadata.nnpdf_metadata["experiment"] != cd2.metadata.nnpdf_metadata["experiment"]
+        assert (
+            cd1.metadata.nnpdf_metadata["experiment"] != cd2.metadata.nnpdf_metadata["experiment"]
+        )
         assert pcd2.experiment != pcd1.experiment
         # but the real experiment is the same
         assert cd1.metadata.experiment == cd2.metadata.experiment
